third_party/WebKit/LayoutTests/inspector/profiler/heap-snapshot-summary-expand-collapse.html

<html>
<head>
  <script src="../../http/tests/inspector/inspector-test.js"></script>
  <script src="heap-snapshot-test.js"></script>
<script>

function test() {
  var instanceCount = 25;
  function createHeapSnapshot() {
    return ProfilerTestRunner.createHeapSnapshot(instanceCount);
  }

  ProfilerTestRunner.runHeapSnapshotTestSuite([function testShowAll(next) {
    ProfilerTestRunner.takeAndOpenSnapshot(createHeapSnapshot, step1);

    function step1() {
      ProfilerTestRunner.switchToView('Summary', step2);
    }

    function step2() {
      ProfilerTestRunner.findAndExpandRow('A', step3);
    }

    function step3(row) {
      row.collapse();
      row.expand();
      var visibleChildren = row.children.filter(c => c._element.classList.contains('revealed'));
      TestRunner.assertEquals(11, visibleChildren.length);
      next();
    }
  }]);
}

</script>
</head>
<body onload="runTest()">
<p>
https://crbug.com/738932
Tests the snapshot view is not empty on repeatitive expand-collapse.
</p>
</body>
</html>
